updated login related components
